requirements.txt
appium-python-client==2.8.0
selenium==4.20.0
selene>=2.0.0b4
pytest==8.2.2
python-dotenv==1.0.1
allure-pytest==2.13.5
requests==2.32.3
jsonschema==4.23.0